user-service-module/user-service-mongo-port/src/main/java/com/kgcorner/topspin/model/UserModel.java
package com.kgcorner.topspin.model;


import org.springframework.data.annotation.Id;

/**
 * Description : Implementation of {@link User} model
 * Author: kumar
 * Created on : 24/11/19
 */

public class UserModel implements User{

    @Id
    private String id;
    private String userName;
    private String name;
    private String email;
    private String contact;
    private String others;
    private boolean active;

    @Override
    public String getId() {
        return id;
    }

    @Override
    public void setId(String id) {
        this.id = id;
    }

    @Override
    public String getUserName() {
        return userName;
    }

    @Override
    public void setUserName(String userName) {
        this.userName = userName;
    }

    @Override
    public String getName() {
        return name;
    }

    @Override
    public void setName(String name) {
        this.name = name;
    }

    @Override
    public String getEmail() {
        return email;
    }

    @Override
    public void setEmail(String email) {
        this.email = email;
    }

    @Override
    public String getContact() {
        return contact;
    }

    public void setContact(String contact) {
        this.contact = contact;
    }

    @Override
    public String getOthers() {
        return others;
    }

    public void setOthers(String others) {
        this.others = others;
    }

    @Override
    public boolean isActive() {
        return active;
    }

    @Override
    public void setActive(boolean active) {
        this.active = active;
    }

}